+++ CHANGELOG.md
@@ -30,6 +30,14 @@ Volume widget:
 
 3D Viewer:
 
+- The line width of skeletons can be adjusted again on platforms other than
+  Linux. This Requires "Volumetric lines" in the "View settings" tab to be
+  enabled (it is by default). Unchecking this option brings back the previous
+  line rendering behavior.
+
+- The new "Focus skeleton" button in the "View" tab will look at the active
+  skeleton's center of mass from the current camea location.
+
 - Volumes can now be smoothed by subdivision. The volume option panel available
   from the View Settings tab now contains a "Subdivide" checkbox.
 
@@ -133,6 +141,8 @@ Miscellaneous:
 
 - 3D viewer: connector restrictions like "show only shared connectors" now
   respect the pre/post visibility toggles in the Selection Table.
+
+- 3D viewer: the initial text scaling for label text is now correctly set again.
 
 - Reconstruction sampler: during interval creation preview, only intervals from
   the currently active domain are now shown.